requirements.txt
tensorflow
numpy
scikit-learn
click
tqdm
PyYAML
PyWavelets
pandas
click